+++ structs.h
@@ -1,5 +1,5 @@
 /** \file structs.h
-	\brief defines the structs 
+	\brief defines the structs
 	\author Gabriele Triglia
 
 	Details.
@@ -11,13 +11,13 @@
 /**
  * \struct Poly
  * \brief strores the coefficients of a polynomial
- * 
+ *
  * coeff[0] + coeff[1]x + coeff[2]x^2 + ...
  */
 
 typedef struct Poly {
-    float *coeffs;
-    int degree;
+	double *coeffs;
+	int degree;
 } poly_s;
 
 #endif /* STRUCTS_H */
